third_party/blink/web_tests/fast/loader/resource-with-idna-deviation-char-in-hostname-iframe-expected.txt

CONSOLE WARNING: The resource at http://faß.test/ contains IDNA Deviation Characters. The hostname for this URL (faß.test) might point to a different IP address after https://chromestatus.com/feature/5105856067141632. Make sure you are using the correct host name.
CONSOLE WARNING: The resource at http://βόλος.test/ contains IDNA Deviation Characters. The hostname for this URL (βόλος.test) might point to a different IP address after https://chromestatus.com/feature/5105856067141632. Make sure you are using the correct host name.
CONSOLE WARNING: The resource at http://ශ්‍රී.test/ contains IDNA Deviation Characters. The hostname for this URL (ශ්‍රී.test) might point to a different IP address after https://chromestatus.com/feature/5105856067141632. Make sure you are using the correct host name.
CONSOLE WARNING: The resource at http://نامه‌ای.test/ contains IDNA Deviation Characters. The hostname for this URL (نامه‌ای.test) might point to a different IP address after https://chromestatus.com/feature/5105856067141632. Make sure you are using the correct host name.
    
